# src/main/resources/generated-spring-api/src/main/java/org/openapitools/api/ApiApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.9.0-SNAPSHOT).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ProductDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-09-10T06:57:22.619286147Z[Etc/UTC]", comments = "Generator version: 7.9.0-SNAPSHOT")
@Validated
@Tag(name = "Product", description = "the Product API")
public interface ApiApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /api/products : Get all products
     *
     * @return Products successfully retrieved (status code 200)
     *         or No products available (status code 204)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getAllProducts",
        summary = "Get all products",
        tags = { "Product" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Products successfully retrieved", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ProductDto.class)))
            }),
            @ApiResponse(responseCode = "204", description = "No products available"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/products",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<ProductDto>> getAllProducts(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"unitPrice\" : 29.99, \"code\" : \"P001\", \"name\" : \"Product 1\", \"description\" : \"This is product 1\", \"stock\" : 100 }, { \"unitPrice\" : 29.99, \"code\" : \"P001\", \"name\" : \"Product 1\", \"description\" : \"This is product 1\", \"stock\" : 100 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/products : Save a new product
     *
     * @param productDto Product to be saved (required)
     * @return Product successfully saved (status code 201)
     *         or Product is not valid (status code 400)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "saveProduct",
        summary = "Save a new product",
        tags = { "Product" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Product successfully saved", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Product is not valid"),
            @ApiResponse(responseCode = "500", description = "Internal server error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/products",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ProductDto> saveProduct(
        @Parameter(name = "ProductDto", description = "Product to be saved", required = true) @Valid @RequestBody ProductDto productDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"unitPrice\" : 29.99, \"code\" : \"P001\", \"name\" : \"Product 1\", \"description\" : \"This is product 1\", \"stock\" : 100 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}